requirements.txt
# blackmirror dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
requests>=2.28.0
beautifulsoup4>=4.11.0
python-nmap>=0.7.1
dnspython>=2.2.0
rich>=12.0.0
pyyaml>=6.0
lxml>=4.9.0
urllib3>=1.26.0 